requirements.txt
streamlit==1.36.0
tensorflow-cpu==2.12.0
numpy
pillow
opencv-python
gdown
